+++ lib/valiot_app_web/schema.ex
@@ -1,7 +1,9 @@
 defmodule ValiotAppWeb.Schema do
   use Absinthe.Schema
+  import Kronky.Payload
   import_types(ValiotAppWeb.Schema.Types)
   import_types(Absinthe.Type.Custom)
+  import_types(Kronky.ValidationMessageTypes)
 
   query do
   end