+++ priv/templates/graphql_schema.exs
@@ -1,8 +1,10 @@
 defmodule ValiotAppWeb.Schema do
   use Absinthe.Schema
+  import Kronky.Payload
   import_types(ValiotAppWeb.Schema.Types)
   import_types(ValiotAppWeb.Schema.ValiotTypes)
   import_types(Absinthe.Type.Custom)
+  import_types(Kronky.ValidationMessageTypes)
 
   query do
     # Users
@@ -77,19 +79,22 @@ defmodule ValiotAppWeb.Schema do
     <% end %><% end %>
   end
 <% end %>
+<%= for {schema, values} <- types do %>
+payload_object(<%= inspect schema |> Inflex.underscore |> String.to_atom %>_payload, <%= inspect schema |> Inflex.underscore |> String.to_atom %>)<% end %>
 
   mutation do
     <%= for {schema, values} <- types do %>
     @desc <%= inspect "Update #{schema |> Inflex.underscore}" %>
-    field :update_<%= schema |> Inflex.underscore %>, type: <%= inspect schema |> Inflex.underscore |> String.to_atom %> do
+    field :update_<%= schema |> Inflex.underscore %>, type: <%= inspect schema |> Inflex.underscore |> String.to_atom %>_payload do
       arg(:id, non_null(:integer))
       arg(<%= inspect schema |> Inflex.underscore |> String.to_atom %>, :update_<%= schema |> Inflex.underscore %>_params)
 
       resolve(&ValiotApp.<%= inspect [Atom.to_string(schema)] |> Module.concat %>Resolver.update/2)
+      middleware &build_payload/2
     end
 
     @desc <%= inspect "Create #{schema |> Inflex.underscore}" %>
-    field :create_<%= schema |> Inflex.underscore %>, type: <%= inspect schema |> Inflex.underscore |> String.to_atom %> do
+    field :create_<%= schema |> Inflex.underscore %>, type: <%= inspect schema |> Inflex.underscore |> String.to_atom %>_payload do
       <%= for {type, attrs} <- values do %><%= case Map.get(attrs, :database) do %>
       <% :normal -> %>arg(<%= inspect type |> Inflex.underscore |> String.to_atom %>, <%= inspect Map.get(attrs, :type) |> Inflex.underscore |> String.to_atom %>)
       <% :has_many -> %>
@@ -98,34 +103,48 @@ defmodule ValiotAppWeb.Schema do
 <% end %><% end %>
 
       resolve(&ValiotApp.<%= inspect [Atom.to_string(schema)] |> Module.concat %>Resolver.create/2)
+      middleware &build_payload/2
     end
 
     @desc <%= inspect "Delete #{schema |> Inflex.underscore}" %>
-    field :delete_<%= schema |> Inflex.underscore %>, type: <%= inspect schema |> Inflex.underscore |> String.to_atom %> do
+    field :delete_<%= schema |> Inflex.underscore %>, type: <%= inspect schema |> Inflex.underscore |> String.to_atom %>_payload do
       arg(:id, non_null(:id))
 
       resolve(&ValiotApp.<%= inspect [Atom.to_string(schema)] |> Module.concat %>Resolver.delete/2)
+      middleware &build_payload/2
     end
 <% end %>end
 
   subscription do
     <%= for {schema, _values} <- types do %>
-    field :update_<%= schema |> Inflex.underscore %>, <%= inspect schema |> Inflex.underscore |> String.to_atom %> do
+    field <%= inspect schema |> Inflex.underscore |> String.to_atom %>_updated, <%= inspect schema |> Inflex.underscore |> String.to_atom %>_payload do
       config(fn _args, _info ->
         {:ok, topic: "*"}
       end)
+      resolve(fn item, _, _ ->
+        {:ok, item}
+      end)
+      middleware &build_payload/2
     end
 
-    field :create_<%= schema |> Inflex.underscore %>, <%= inspect schema |> Inflex.underscore |> String.to_atom %> do
+    field <%= inspect schema |> Inflex.underscore |> String.to_atom %>_created, <%= inspect schema |> Inflex.underscore |> String.to_atom %>_payload do
       config(fn _args, _info ->
         {:ok, topic: "*"}
       end)
+      resolve(fn item, _, _ ->
+        {:ok, item}
+      end)
+      middleware &build_payload/2
     end
 
-    field :delete_<%= schema |> Inflex.underscore %>, <%= inspect schema |> Inflex.underscore |> String.to_atom %> do
+    field <%= inspect schema |> Inflex.underscore |> String.to_atom %>_deleted, <%= inspect schema |> Inflex.underscore |> String.to_atom %>_payload do
       config(fn _args, _info ->
         {:ok, topic: "*"}
       end)
+      resolve(fn item, _, _ ->
+        {:ok, item}
+      end)
+      middleware &build_payload/2
     end
   <% end %>
   end