+++ priv/templates/graphql_schema.exs
@@ -123,9 +123,7 @@ payload_object(<%= inspect schema |> Inflex.underscore |> String.to_atom %>_payl
       config(fn _args, _info ->
         {:ok, topic: "*"}
       end)
-      resolve(fn item, _, _ ->
-        {:ok, item}
-      end)
+      resolve(&ValiotApp.SubscriptionResolver.subscribe/3)
       middleware &build_payload/2
     end
 
@@ -133,9 +131,7 @@ payload_object(<%= inspect schema |> Inflex.underscore |> String.to_atom %>_payl
       config(fn _args, _info ->
         {:ok, topic: "*"}
       end)
-      resolve(fn item, _, _ ->
-        {:ok, item}
-      end)
+      resolve(&ValiotApp.SubscriptionResolver.subscribe/3)
       middleware &build_payload/2
     end
 
@@ -143,9 +139,7 @@ payload_object(<%= inspect schema |> Inflex.underscore |> String.to_atom %>_payl
       config(fn _args, _info ->
         {:ok, topic: "*"}
       end)
-      resolve(fn item, _, _ ->
-        {:ok, item}
-      end)
+      resolve(&ValiotApp.SubscriptionResolver.subscribe/3)
       middleware &build_payload/2
     end
   <% end %>