+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Master
 - [New feature] Added kronky for better error responses on mutations
+- [New feature] Changed subscriptions names
+- [New feature] Add unique field support on generator
 # 0.1.0
 - [Bug Fix] Enum with multiple words works now
 - [Bug Fix] Queries for inserted_at and created_at work now