+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Master
+- [New feature] Added kronky for better error responses on mutations
+- [New feature] Changed subscriptions names
 - [New feature] Add unique field support on generator
 # 0.1.0
 - [Bug Fix] Enum with multiple words works now