+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - [New feature] Changed subscriptions names
 - [New feature] Add unique field support on generator
 - [New feature] Add filter by list of ids
+- [Bug Fix] Subscriptions now require authorization
+
 # 0.1.0
 - [Bug Fix] Enum with multiple words works now
 - [Bug Fix] Queries for inserted_at and created_at work now