+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Master
+
+# 0.2.0
 - [New feature] Added kronky for better error responses on mutations
 - [New feature] Changed subscriptions names
 - [New feature] Add unique field support on generator
 - [New feature] Add filter by list of ids
 - [Bug Fix] Subscriptions now require authorization
+- [Maintance] Update dependencies, elixir, and erlang
 
 # 0.1.0
 - [Bug Fix] Enum with multiple words works now